flatten the module 'recognizer'
